+++ src/main/java/org/stellar/sdk/requests/sorobanrpc/SendTransactionRequest.java
@@ -5,7 +5,7 @@ import lombok.Value;
 /**
  * Request for JSON-RPC method sendTransaction.
  *
- * @see <a href="https://soroban.stellar.org/api/methods/sendTransaction#parameters"
+ * @see <a href="https://developers.stellar.org/docs/data/rpc/api-reference/methods/sendTransaction"
  *     target="_blank">sendTransaction documentation</a>
  */
 @Value