+++ src/main/java/org/stellar/sdk/responses/sorobanrpc/SendTransactionResponse.java
@@ -10,7 +10,7 @@ import org.stellar.sdk.xdr.TransactionResult;
 /**
  * Response for JSON-RPC method sendTransaction.
  *
- * @see <a href="https://soroban.stellar.org/api/methods/sendTransaction#returns"
+ * @see <a href="https://developers.stellar.org/docs/data/rpc/api-reference/methods/sendTransaction"
  *     target="_blank">sendTransaction documentation</a>
  */
 @Value